The LastN set for a participant must not include the local participant.
When a participant clicks on his/hers thumbnail Jitsi meet sends to the
bridge the pinned endpoint event containing the local jid. This is
expected and required to update the internal state of the bridge.

The problem is that when the bridge receives the event, it (wrongfully)
puts the local endpoint in the lastN set and removes one of the remote
peers from the lastN set and stops the video of the remote peer that is
removed from the lastN set. Then jitsi meet receives event for lastN
changed with the local jid in it.