Fixes an issue where a simulcast layer can't be decrypted after a while.
For SRTP replay protection the webrtc.org implementation uses a replay
database with extended range, using a rollover counter (ROC) which
counts the number of times the RTP sequence number carried in the RTP
packet has rolled over.

In this way, the ROC extends the 16-bit RTP sequence number to a 48-bit
“SRTP packet index”. The ROC is not be explicitly exchanged between the
SRTP endpoints because in all practical situations a rollover of the RTP
sequence number can be detected unless 2^15 consecutive RTP packets are
lost.

For every 0x800 (2048) dropped packets (at most), send 8 packets so that
the receiving endpoint can update its ROC.