Fixes the initial simulcast stream selection.
When data channel connection establishment is fast and if the selected
endpoint data channel message is sent early enough then when one user
(re)joins a call, it's getting a good quality stream from the participant
that was already in the room. But the user that was already in the room,
gets a low quality stream from the newly joined client. So effectively,
in a 1:1 session, one user always ends up with a low quality stream of
the remote... The user that joined first.