Revert 5c577d2
TcpHarvester binds to specific local addresses, Jetty binds to all/any
local addresses. Consequently, Jetty will fail to bind if TcpHarvester
binds first and Jetty is supposed to share a port with TcpHarvester. To
resolve this failure, initialize TcpHarvester as late as possible and
leave Jetty to bind at startup.