Implements performance optimizations around simulcast.
Moves the SimulcastReceiver early on in the transform chain (before
SRTP) and implements logic to keep track of which of the simulcast
streams are in use (by a SimulcastSender).

Marks packets from currently unused streams with the Buffer.DISCARD
flag, which causes them to not be decrypted and to be discarded before
they pass the RTPTranslator (before copies for each receiver are made).

Refactors SimulcastSender initialization, so that it doesn't depend
on packets flowing through the chain. It happens based on signalling.