Updates the collect-dump-logs.sh
We want to make sure we grab the java heap dump, even if the process
appears to be running.  If the JVB has crashed (with OOM, for example),
the JVM might be configured to take a heap dump. During that time, the
process will appear running, but the JVM won't allow us to connect to it
and take a heap/thread dump.