feat: detect when Channel was never active
It can happen that conference participant will fail to send
the initial response or ICE fails immediately and we want
to notify other participants about this failure.

Endpoint instances are stored as inactive until the Conference
expire, because the person may still remain in the conference
room for unknown amount of time. The status monitor needs
to know about it to notify new participants who have missed
the ad-hoc notification.