Modifying code to comply with coding conventions.
For the record: this change hurts readability of the code, and therefor defeats the
purpose of having a coding convention in the first place.

Of note: the coding convention used here[1]  links to a standard that is explicitly
defined to be no longer maintained since April 20, 1999. Which, coincidently, was when
some 10 years after terminals that had a 80-character width limit stopped being fahionable.

It's good to have consistency where needed, but consistency should never be a goal in itself.

1: https://jitsi.org/Documentation/CodeConvention