Component should expose its videobridge instance
This is a fix for https://github.com/jitsi/jitsi-videobridge/issues/335

The Component implementation that wraps the Videobridge instance should expose that Videobridge instance publicly, instead of privately.

Obtaining the Videobridge instance can be achieved with other data that is publicly exposed by the component (notably getBundleContext() and ServiceUtils2.getService), and the implementation already contains a private method that exposes the instance. This method can be made publicly accessible, which would remove the need to duplicate its implementation.