fix(ls-hack): Supports dynamic PT for VP8.
Chrome 57 changed the default payload types (e.g. previously it was 100
for VP8 and now it's 96). We're still using the old payload types and
now Chrome seems to not register the associated payload type for VP8 RTX
which breaks retransmissions. The workaround is to use the new payload
type defaults that Chrome 57 is using, so the LS hack needs to
dynamically discover the VP8 payload type and not always assume it's
100.