Native libraries should be attached as Maven artifacts.
To be able to re-use them, the various native libraries that ship with the
Videobridge should be added as a Maven artifact. Using the change in this
commit, a Maven project could reference the native libraries by defining
a dependency like this:

    <dependency>
        <groupId>org.jitsi</groupId>
        <artifactId>jitsi-videobridge</artifactId>
        <version>1.0-SNAPSHOT</version>
        <classifier>libraries-native-linux-32</classifier>
        <scope>runtime</scope>
    </dependency>

The wording of the assembly identifier (and thus, the classifier,
and thus, the resulting .jar name) was worded to be compatible
with the checkNatives() implemenation, that checks for files
matching patterns like these:

- ``*-native-linux-32.jar``
- ``*-native-linux-64.jar``
- ``*-native-windows-32.jar``
- ``*-native-windows-64.jar``
- ``*-native-macosx.jar``

Unlike the pre-existing assemblies, these new assemblies do not generate
'distributables', but 'artifacts' instead. Therefor, ${assembly.skipAssembly}
should not apply to these new ones. Also: the file size for the libraries
is limited, unlike the filesize of the full, per-architecture distributables.