fix: Do not attempt to rewrite the TL0PICIDX of non-scalable streams.
Anything that's not a desktop Chrome browser (e.g. FF, mobile app) does
not support simulcast atm. Attempting to rewrite the TL0PICIDX of these
non-scalable streams results in the logs being spammed with the message
"Failed to set the VP8 TL0PICIDX". This commit fixes this by not
attempting to rewrite the TL0PICIDX of non-scalable streams.