feat: Implements bitrate allocation for multiple tracks per endpoint.
Other than being a useful feature, this resolves an issue where the JVB
does not forward a FF stream in a multi-party call. FF signals multiple
local tracks (because of unified plan). The current JVB code assumes a
single track per endpoint, so it only shapes the traffic for the first
video track. So, if the send track from FF happens to be in the end of
the tracks list, it never gets forwarded.