fix: release resources when shutting down. (#663)
* fix: release resources when shutting down.

When the bridge is being closed, resources should be released explicitly. This commit
closes down harvesters, which in turn unbinds from the network interface.

* fix: release resources when shutting down.

When the bridge is being closed, resources should be released explicitly. This commit
clears system properties that were set during initialization.

* fix: release resources when shutting down.

When shutting down, reset healty flag to it's initial state and null the harvester fields.

* fix: release resources when shutting down.

Add comment that explains why system properties are being cleared.