more work on the packet pumping through the pipeline
added some more stats/tweaked the way some are calculated.  tried to
verify some basic assumptions (like the fact that the blocking queue
won't be a bottle neck--it won't be).  throughput times here (with very
little work being done) all seem to be in line: they can't look "low"
but that appears to be largely based on the pace at which things are written.