plumb in the rtpencodings
send through the created rtpencodings to the nlj stuff so they can be
updated as packets come in.  this makes sure that the same rtpencodings
we update are the ones that the bitratecontroller/simulcastcontroller/etc. chain sees, so that chain is (hopefully) happy.  video is flowing, but can't test more right now since datachannel isn't working.