don't create/take an executor in dtlsclientstack
we only used the executor one time: in the connect call.  instead it
makes more sense to make the client stack simpler and just have connect
execute directly in the calling thread: the caller can decide to post
the connect call to a an executor/in a new thread/whatever it wants.