ref(SctpConnection): let the client open WebRTC data channel
Modifies the logic to use WebRTC data channels opened by the client. The
bridge will still open it's channel if the client doesn't do that early.
In case multiple channels are open at the same time the bridge will try
to write on the client's channel and listen on all of them.

This commit also contains a fix for ordering of the SCTP processing, so
that any notifications about data/packets will come out of
the SctpConnection only after it's underlying SCTP socket is writeable.
It fixes a case where the bridge tries to respond with an ack to
MSG_OPEN_CHANNEL when the socket is not ready yet.