pass transport when creating dtls stack, not on connect
the plan is to use this model for other protocol stacks (as of now sctp
and data channel).  the stack gets created with a transport it can use
for receiving and sending.  the stack will then provide a transport for
receiving and sending app packets (i.e. non-control packets, which the stack itself will
handle) on top of that protocol.