fix a bug where we mistakenly think ice is connected
we use the ice4j helper IceProcessingState.isEstablished to check if ice
is connected, but this can lead to false positives as it just checks
whether the state is COMPLETED or TERMINATED.  however, some cases (like
i'm seeing with health checks where ice fails but doesn't go to failed?)
go directly from RUNNING to TERMINATED.  isEstablished returns true
there, so we thought things were connected.  for now, just check once we
hit completed, but i'm not sure if this is reliable?