feat(ComponentImpl): use StripedExecutorService
Uses StripedExecutorService to fix the order in which the packets are
processed by the XMPP component. Packets related to one conference must
be processed in the order they were received. That is achieved by
producing a stripe for a conference ID extracted from a Colibri packet.
See docs of the ComponentImpl.getStripeForPacket method for more info.
See also: https://www.javaspecialists.eu/archive/Issue206.html