fix(EndpointConnectionStatus): leaking expired Endpoints (#739)
The EndpointConnectionStatus stores 'inactive' status even for Endpoints
that have expired in order to show the disconnected status in Jitsi-Meet
after the channels could have been expired by Jicofo, but the person
stayed in the room.

The current cleanup logic will remove it only after the parent
conference has expired. That is causing problem in case the channels are
expired and re-created for Endpoints within the same conference. This
commit adds an extra check for eventual Endpoint instance replacement.