tag video packets with a bitrate snapshot
this brings in modified/slimmed down versions of MediaStreamTrackDesc
and RTPEncodingDesc and adds an API to see the MediaStreamTracks on the
transceiver.  These are used to look up the RTPEncodingDesc for a given
packet so we can calculate the current bitrate of each RTPEncoding
(which will be used by egress logic to decide whether or not packets
shoudl be accepted/forwarded).