fix a bug when creating an adaptivetrackprojectioncontext
adaptivetrackprojection no longer reaches into a central location to check for signalled
payload types, it waits to be notified.  i saw an instance where we got
a packet before getting the payload type signalled and therefore created
a projection which couldn't work (vp8 or generic) because we didn't have
a valid payload type.  this change makes sure we don't create the
context until we know the payload type.

it'd be great to guarantee this couldn't happen, so maybe this is worth
revisiting to see if it can be improved.