change the way outgoing packets are handled
instead of adding them to a queue, invoke a handler instead.  this is so
that the external entity (e.g. jvb) can use a PacketQueue on its side
for handling the packets.

NOTE: i don't necessarily think this is ideal, but the handler has to be
passed at the time the PacketQueue is created, and we want that handler
to be in jvb.