better trigger BitrateController updates
this change changes the API of BitrateController to better reflect the
events that can drive updates and only require what is needed to be
passed in (and caching the rest to be re-used where needed).  this
change also adds a new event to fire when an endpoint's
mediastreamtracks are set so we can trigger bitratecontroller updates.