fix: Partially handles screen-sharing simulcast.
This is the first PR from a (hopefully short) series of PRs that will
implement proper support for simulcast for screen-sharing.

When VP8-based screen-sharing is activated Chrome is sending 2 streams
with the same resolution but with different bitrates and different
temporal layer bitrate allocations. This is different from what the
bridge expects, which is 3 streams with _different_ resolutions. The
different resolutions imply a quality order enabling the bridge to
switch from low quality to high quality and vica-versa.

Since we can no longer rely on the streams having different resolutions,
there's no longer an implicit order that the bridge can detect
on-the-fly, so now we take this information from the signaling.