fix a bug in calculating TCC deltas
previously we used the full reference timestamp to calculate the deltas,
but when we serialize the data we lose precision on that value (it's
turned into a multiple of 64ms, so we lose the lower 6 bits of data).
this means that when the far side reconstructs the timestamps, they get
a different result (we calculated the deltas against a full-precision
reference time, but the far side rebuilds the timestamps using a
lower-precision reference time), this gets detected as delay/jitter and
causes bwe issues.