break a possible deadlock around sctp messages
previously we had a continuous chain (no posting to
another thread) from an sctp packet coming in, the handler being
invoked, going through to the data channel handler, bubbling up to
endpointmessagetransport and then being sent out.  so the lock in
SctpSocket#onSctpIn is never released all the way to sending it out, so
when another endpoint tries to send to IT: they both hold their own
locks in SctpSocket#onSctpIn and are trying to get the same lock (of the
other side) to send out.