change the way we create PaddingVideoPacket
before we pulled a buffer from the pool in the ctor and immediately
invoked super with it (because we had to).  this means that any code
above which may do any parsing (like in the header) could be dealing
with garbage.   i got bit by this again and i think the state it was in
was just bad and very easy to make a mistake with.