track Packet delay stats (#95)
* add packet delay tracker to rtp sender

* use a delay of -1 in the timeline to denote there were no events

also clean up the toString() if the timeline isn't enabled

* clean up print if timeline is disabled

* move DelayStats out, change fields to atomic and make the instance static on RtpSender

this will aggregate the stats across all RtpSenders.  this also means
that it will no longer be printed as part of the pipeline stats, but the
idea is that it will be queried from elsewhere and reported somewhere
(e.g. the debug api)

* use longaddr instead of atomics

* undo accidental change

* fix linter errors

* create generic delaystats class

* rename variable

* remove e2e packet delay from rtp sender

(will be moving this to outgoing dtls transport send in jvb, which is
more accurate)

* add helper to convert DelayStats to JSON

* fix linter errors