unthrottle bitrate allocation updates when the bandwidth increases
Right now we require a 15% (configurable) change in the bwe in order to
update the bitrate allocation. This is an ugly hack to prevent too many
resolution/UI changes in case the bridge produces too low bandwdith
estimate, at the risk of clogging the receiver's pipe.

The code prevented not only downscaling, but also upscaling, which is
problematic: Suppose that the target bitrate is 2.5Mbps, and that the
last bitrate allocation was performed with a 2.4Mbps bandwidth esitmate.
The bridge keeps probing and, suppose that, eventually the bandwidth
estimate reaches 2.6Mbps, which is plenty to accomodate the target
bitrat; but the minimum bandwidth estimate that would trigger a new
bitrate allocation is 2.4Mbps + 2.4Mbps * 15% = 2.76Mbps.