remove periodic feedback features of tcc
we're not using it and it makes testing tcc generator more difficult
(more dependencies to satisfy).

it was ported from the chrome code but when i tried i got slower
rampups.  it's possible we may need to use a hybrid approach: non-period
for a time period and then switching to periodic?