gave this a shot, but i don't think it'll  give me what i want:
the idea was that then i’d have a single, top-level place to gate access
to colibri as well.  unfortunately this didn’t work: this work is
‘pre-matching’ from a filter perspective and, if i mark the filter with
@PreMatching, then the request isn’t mapped to a resource class yet, so
we can’t check for the annotation.  (note that this scheme would also
mean that i couldn’t add the whole package, but instead only the
top-level  classes  like  Colibri—not terrible but the package adding is
nice.