use new config in videobridge
NOTE that i had to add surefire plugin here to set 'reuseForks' to false
because we set a system property for a unit test in bridgeshutdowntest
and it wasn't getting reflected properly when running the tests in
parallel.  i don't think this is the worst workaround, but maybe there
is a better way to to load the config/override values for tests.