refactor the way properties are retrieved
this makes a lot of changes to the underlying helper classes used to
define a new configuration property.  largely this is just some
cleanup/reorganization/improvement to make the definition of a new
property easier and more flexible (for example, allow for a
transformation of a retrieved value to be applied).