add jvbconfig.init to focuscontroltest
this is needed because the tests don't run Main, where the configuration
is normally initialized.  i looked at moving this code to the
start method of Videobridge, but that won't work either as this isn't
run before main and main needs the config as well.  i could have it in
both places, but worried about double-initializing config during normal
run as that effectively re-loads the config.