no longer require an explicit loading of jvb config
i had previously made this explicit mainly to fix some issues with unit
tests, but also because i thought it would be better to make it clearer
when the config files were being loaded.  unfortunately, this caused
issues with some flows (tests, mainly, but also loading videobridge as
an osgi service).  the config load can't be put in the videbridge osgi
loader, as Main doesn't use that before loading some config it wants.
this also causes issues with jicofo which loads jvb as a library in its
tests.  anyway, the test config loading has changed a bit and works fine
with this method (implicit initialization of config) so we'll try this
and adapt in the future if need be.