Config phase 1 (#939)
* add jvb new config helper classes

* migrate health interval to new property type

* add config readme

* dynamically retrieve the confiig inside the suppliiers

this is important, because we replace the config instances on config
reload, so we need to make sure we get the most up-to-date instance

* initialize the jvb config in main

* fix accidental not-found-strategy for health interval property

* add unit test for  healthintervalproperty

* config readme tweak

* change where we log a found legacy config

with this change we no longer log an empty legacy config if one wasn't
actually found

* reorganize some logs around config loading, throw if videobridge config isn't found

* config readme tweaks

* more config readme tweaks

* add traiiling newline to reference.conf

* add jvbconfig.init to focuscontroltest

this is needed because the tests don't run Main, where the configuration
is normally initialized.  i looked at moving this code to the
start method of Videobridge, but that won't work either as this isn't
run before main and main needs the config as well.  i could have it in
both places, but worried about double-initializing config during normal
run as that effectively re-loads the config.

* always add at least an empty new config in prop tests

this is needed since we now throw if the 'videobridge' config block
isn't found

* init config in bridgeshutdowntest

* no longer require an explicit loading of jvb config

i had previously made this explicit mainly to fix some issues with unit
tests, but also because i thought it would be better to make it clearer
when the config files were being loaded.  unfortunately, this caused
issues with some flows (tests, mainly, but also loading videobridge as
an osgi service).  the config load can't be put in the videbridge osgi
loader, as Main doesn't use that before loading some config it wants.
this also causes issues with jicofo which loads jvb as a library in its
tests.  anyway, the test config loading has changed a bit and works fine
with this method (implicit initialization of config) so we'll try this
and adapt in the future if need be.

* add config validation

* update jitsi-utils

* minor version bump