move the configuration of config source suppliers to another file (#967)
* move the configuration of config source suppliers to another file

i noticed this when working on another PR when i tried setting the
command-line args supplier to include the args given in main.  the
problem was that, since JvbConfig loaded the config in a static block,
any access on JvbConfig would trigger it--this includes setting a
supplier.  for the command-line args case, this means that when you
called JvbConfig.commandLineArgsSupplier = ..., the config would load
and log that no command line args were present (since the assignment was
not yet complete) which was misleading.  it also had the annoying
behavior of dumping any default-loaded config during tests when the
suppliers were overriden.

by separating that config out, we can manipulate the config source
suppliers without triggering a loading of the config, which makes more
sense.

* formatting/qualify config source suppliers