remove the explicit 'enabled' prop for client configs
since the old code didn't have this, we run into a problem when we pull
the configs themselves from the old config, as there's no 'enabled'
there so we fallback to the 'enabled' value in the new config.  i have
thoughts on how we could make a property like 'enabled' here depend on
_where_ another config value was pulled from, but that would require
some bigger changes.  'enabled' isn't critical here, so we can come back
to how to restore it (if we want) later.