Propagate selected/pinned events across all bridges. (#945)
This work is a step towards interbridge last-n. By propagating
selected/pinned events across all bridges, every bridge in a particular
call is aware of the selected and pinned participants in every other
bridge, hence making it possible to determine whether a participant is
globally off-stage or on-stage.

If the participant is globally on-stage, her media needs to be
broadcasted across all bridges (in addition to the any participants that
are in the last-n set, of course). Otherwise (i.e. the participant is
off-stage), the bridge can notify the participant to suspend their SD/HD
layers.