Ice transport refactor (#1104)
Refactors the ICE transport logic into a standalone network "layer".

This layer no longer tries to fit into the "pipeline" (node) model of other parts of the packet flow, but works more like a network layer with callbacks for incoming data and methods for sending data out.  It says the stage for a more direct path for SRTP packets (which don't need to go through any "DTLS" layer) and more clearly distinguishes the paths for SRTP and DTLS packets.

This is the first part of a refactor of the transport layers to shift to this callback-based model.