JVB http (ws) api v1 (#1268)
v1 of the JVB http (websocket, in this case) API. It is implemented as 3 artifacts (independent of the JVB artifact) living in the JVB repo:

jvb-api-common: common code used by both jvb-api-client and jvb-api-server.
jvb-api-client: the api "client" code. client here refers to the entity controlling the jvb.
jvb-api-server: the api "server" code. this is what jvb itself will integrate with to handle requests.
v1 of this API defines 2 methods:

sendIqAndGetReply(IQ): IQ which synchronously sends an IQ and returns the reply
sendAndForget(IQ): which sends an IQ but does not wait for the response
Basically, v1 does nothing but change the transport over which existing COLIBRI messages are sent. This implementation uses websockets in order to preserve the ordering of the requests (which COLIBRI requires).

It also restructures the pom structure a bit: `pom.xml` is now a parent pom for both the jvb-api module and the jvb module.

The code for integrating the API into the bridge is not included here.