ref: Unifies pinned, selected and receiver video constraints. (#1250)
This PR introduces video constraints a which unify the last-n, pinned, selected and max receiver resolution messages to the bridge, all while providing full backwards compatibility with existing client side code. A nice side effect of this is less ninjas in tile-view. It fixes issue #147
and it also addresses pull request #1249.