Fix DataChannel SCTP packets arriving out-of-order
When opening a DataChannel SCTP connection on a heavily loaded machine, the logs often indicate:
> SEVERE: Could not find data channel for sid <sid>
This indicates that data packets arrived before the OpenChannelMessage was processed.

The cause of this issue is that incoming packet processing was scheduled on independent IO threads.
As such, they would be handled in any particular order on a loaded machine.
While JVB could recover from this condition for future packets, initial packets would get lost.
I have observed examples where this impacted initial Last-N selection, for which the command is sent from Jitsi Meet upon connection opened.

This change ensures that incoming SCTP packets are processed in the exact order that they arrived in.