ref: Remove "stats transports".
Refactors the way stat transports are configured. Instead of having the
stats manager maintain a list of transports to enable, each module that
uses this functionality (XmppConnection, CallstatsService) calls into
the stats manager and adds its transport. The code which parses the
configuration is still present and used for backward compatibility.