fix: Increase the T3 threshold to 1500.
We observe RTP packets with size >1240 (the previous threshold). In this
case we always allocate a new buffer (since the T3 buffers are not
guaranteed to satisfy this size), and later return it to T3. This
ends up blowing up the T3 pool, causing GC overhead. To prevent this
we don't return buffers >T3 to the pool anymore.