fix: Delay shutdown in graceful mode. (#1756)
* fix: Delay shutdown in graceful mode.

When a graceful shutdown is requested while there are no conferences on
the bridge it would shutdown immediately, without updating its presence
with the "graceful-shutdown" flag. This resulted in jicofo wrongly
detecting the bridge as having gone down non-gracefully.

This adds a delay to give time for presence to be updated.