fix: Consider all spatial layers for VP9 screensharing
With VP9 the lowest spatial/quality layer doesn't always have the lowest
bitrate. In this case we can't match the resolution requested by the
client, but there's no good reason to limit to the lowest
spatial/quality layer.

Note: the jitsi-meet client requests to receive maxHeight=2160 in the
case of screensharing.