feat: Suppress UNKNOWN_ENDPOINT errors for some requests
When the request is just trying to expire and endpoint or relay that is
missing (presumably already expired) we don't need to throw an error.
Same if the request only updates the force-mute state of an endpoint
(the only other operation that is currently performed in a batch by
jicofo).

Other requests, which create an endpoint/relay or update the
media/sources/transport of an endpoint still throw UNKNOWN_ENDPOINT
erorrs. These requests are now required to reference a single endpoint
or relay (to simplify error handling).