ref(colibri2 session participant re-invite): specific colibri2 unknown endpoint error (#1999)
* Added functionality to send back UNKNOWN_ENDPOINT Colibri2Error when there is no such endpoint. This addition reason will help jicofo to recognise this colibri2 error and re-invite participant instead of invalidating whole bridge and re-inviting all participants.

* ref: Move UnknownEndpointException definition.

* Update jitsi-xmpp-extensions.

* feat: Return feature_not_implemented for colibri2 requests

that use features known not to be properly supported by jitsi-videobridge
(and not used in jicofo). This is mostly to help spot problems early if
developing a different colibri2 client (not jicofo).

* feat: Suppress UNKNOWN_ENDPOINT errors for some requests

When the request is just trying to expire and endpoint or relay that is
missing (presumably already expired) we don't need to throw an error.
Same if the request only updates the force-mute state of an endpoint
(the only other operation that is currently performed in a batch by
jicofo).

Other requests, which create an endpoint/relay or update the
media/sources/transport of an endpoint still throw UNKNOWN_ENDPOINT
erorrs. These requests are now required to reference a single endpoint
or relay (to simplify error handling).

---------

Co-authored-by: Boris Grozev <boris@jitsi.org>