Add a Java option to work around a major performance regression in JDK 17 (#2110)
There was a rewrite of the Java networking stack in Java 13 which made
it impossible to concurrently send UDP from multiple threads on the
same DatagramSocket.  See https://bugs.openjdk.org/browse/JDK-8303616.
This Java option causes Java to use its old DatagramSocket implementation.

Note that unfortunately this workaround does not work in Java 18 and later.