fix: Remove mappings when an endpoint expires.
With SSRC rewriting source mappings are not removed when an endpoint
expires. This results in duplicate entries in the other endpoints'
mappings when an endpoint does an ICE restart (i.e. is recreated
with the same ID). Subsequently, if an endpoint's WS reconnects,
we signal all mapping including duplicates which may cause a
failure to playback/render some sources.