fix: Sources with type NONE should not have effective constraints.
When a source has video type NONE or DISABLED, it has been signaled but
is not transmitted. For the purpose of generating effective constraints,
it should be as if the source was not present.

A change of effective constraints is the trigger for creating a new
mapping for a receiver. If a receiver has effective constraints for a
source with type NONE, and that source is later enabled, the effective
constraints don't change, and the mapping is not created. This should
fix #2288.