Merge from WebRTC: Add posibility to scale max_allocated bitrate when deciding to skip probe.
Add field trial parameter for setting scale factor of max allocated bitrate used for deciding when to skip probing.
Currently, a factor of 2 is used in most places for max allocated bitrate but not if the field trial skip_estimate_larger_than_fraction_of_max is used.
The purpose of this new field parameter is to be able to harmonize and always use the same factor.

WebRTC commit 17642c0db96aa6ee03af4319fc7895bb314e89a9